+++ src/domains/admin/admin-dashboard/admin-dashboard-sites/constants/site-link-data.ts
@@ -1,0 +1,33 @@
+export const SITE_LINK_DATA = [
+  {
+    id: 1,
+    link: "https://github.com/BorussiaDortmundKorea-Scored/bdks-fe",
+    name: "github",
+  },
+  {
+    id: 2,
+    link: "https://supabase.com/dashboard/project/abxgeyabzwzrorecsjcd",
+    name: "supabase",
+  },
+
+  {
+    id: 3,
+    link: "https://vercel.com/youngducks-projects/bdks-fe",
+    name: "vercel",
+  },
+  {
+    id: 4,
+    link: "",
+    name: "sentry",
+  },
+  {
+    id: 5,
+    link: "",
+    name: "notion",
+  },
+  {
+    id: 6,
+    link: "",
+    name: "패치노트",
+  },
+];