+++ src/domains/admin/admin-dashboard/admin-dashboard-user-count/api/react-query-api/admin-dashboard-user-count-query-keys.ts
@@ -1,0 +1,3 @@
+export const ADMIN_DASHBOARD_USER_COUNT_QUERY_KEYS = {
+  USER_TOTAL_AND_MONTHLY_PERCENT: "admin-dashboard-user-count-total-and-monthly-percent",
+} as const;