+++ src/shared/mocks/handlers/handlers.ts
@@ -1,13 +1,16 @@
+import { AdminDashboardUserCountHandlers } from "@admin/admin-dashboard/admin-dashboard-user-count/mocks/admin-dashboard-user-count-handler";
+
 import { AnimalHandlers } from "@animals/mocks/animal-handlers";
+
+import { MatchesHistoryHandlers } from "@matches/matches-history/mocks/matches-history-handler";
 
 import { PlayersDBHandlers } from "@players/players-db/mocks/players-db-handler";
 import { PlayersRatingRotatorHandlers } from "@players/players-rating-rotator/mocks/player-rating-rotator-handlers";
 
-import { MatchesHistoryHandlers } from "@matches/matches-history/mocks/matches-history-handler";
-
 export const handlers = [
+  ...AdminDashboardUserCountHandlers,
   ...AnimalHandlers,
-  ...PlayersRatingRotatorHandlers,
-  ...PlayersDBHandlers,
   ...MatchesHistoryHandlers,
+  ...PlayersDBHandlers,
+  ...PlayersRatingRotatorHandlers,
 ];